requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
motor>=3.6.0
pydantic-settings>=2.4.0
firebase-admin>=6.5.0
crewai>=0.86.0
